VQ-431:Unit test and Documentation on Premium Rates

+++ src/app/features/gis/components/setups/components/premium-rate/premium-rate/premium-rate.component.ts
@@ -6,7 +6,7 @@ import { SubclassesService } from '../../../services/subclasses/subclasses.servi
 import { BinderService } from '../../../services/binder/binder.service';
 import { Table } from 'primeng/table';
 import { NgxSpinnerService } from 'ngx-spinner';
-import { Logger } from 'src/app/shared/shared.module';
+import {Logger} from '../../../../../../../shared/shared.module'
 
 const log = new Logger('PremiumRateComponent');
 
@@ -17,7 +17,7 @@ const log = new Logger('PremiumRateComponent');
 })
 
 /**
- * This Angular class, `PremiumRateComponent`, manages the functionality related to insurance premium rates.
+ * This Angular class, `PremiumRateComponent`, manages the functionality related to premium rates.
  *  It handles the selection of subclasses, sections, binders, and premium rates, 
  * allowing users to view, create, update, and delete premium rate data. 
  */
@@ -118,13 +118,23 @@ export class PremiumRateComponent implements OnInit {
     // this.loadAllPremiums(code);
     
   }
-  
+  /**
+ * Filters the list of subclasses based on a search input.
+ * It takes an event containing the user's input, converts it to uppercase for case-insensitive matching,
+ * and filters the 'subClassList' to include only those elements whose 'description' property includes the search value. 
+ * @param event - The event object containing the user's input.
+ */
   filterSubclass(event: any) {
     const searchValue = (event.target.value).toUpperCase();
     this.filteredSubclass = this.subClassList.filter((el) => el.description.includes(searchValue));
     this.cdr.detectChanges();
   }
-
+/**
+ * Loads and updates sections for a given subclass code.
+ * Retrieves section data via an HTTP request, updates 'sectionList'
+ * and 'filteredSection,' and triggers UI updates.
+ * @param code - The subclass code for which sections are loaded.
+ */
   loadAllSections(code:any){
     return this.subclassService.getSubclassSectionBySCode(code).subscribe(data=>{
       this.sectionList=data;
@@ -133,11 +143,24 @@ export class PremiumRateComponent implements OnInit {
       this.cdr.detectChanges();
     })
   }
+
+  /**
+ * Filters the list of sections based on a search input.
+ * It takes an event containing the user's input, converts it to uppercase for case-insensitive matching,
+ * and filters the 'sectionList' to include only those elements whose 'sectionShortDescription' property includes the search value. 
+ * @param event - The event object containing the user's input.
+ */
   filterSection(event: any) {
     const searchValue = (event.target.value).toUpperCase();
     this.filteredSection = this.sectionList.filter((el) => el.sectionShortDescription.includes(searchValue));
     this.cdr.detectChanges();
   }
+  /**
+ * Updates the selected section and its code based on provided data.
+ * Typically used when a user selects a section.
+ *
+ * @param data - The selected section data.
+ */
   loadSections(data:any){
     this.selectedSection=data;
     // log.debug(this.selectedSection,'This is a section')
@@ -145,6 +168,13 @@ export class PremiumRateComponent implements OnInit {
     // this.loadAllPremiums(selectedItemCode);
    
   }
+  /**
+ * Loads and updates the list of binders for a specific subclass.
+ * Retrieves binder data via an HTTP request, updates 'binderList' and 'binderListDetails,'
+ * and filters 'selectedBinderList' to include only binders matching the provided subclass code.
+ *
+ * @param code - The subclass code for which binders are loaded.
+ */
   loadAllBinders(code:any){
     this.binderService.getAllBinders().subscribe(data=>{
       this.binderList=data;
@@ -155,16 +185,34 @@ export class PremiumRateComponent implements OnInit {
 
     })
   }
+  /**
+ * Sets the selected binder code based on the provided code.
+ * This method is typically called when a user selects a binder, and it updates the 'selectedBinder'
+ * property with the provided code.
+ * @param code - The code of the selected binder.
+ */
   selectedBinderCode(code: any){
     this.selectedBinder=code;
     // log.debug("The selected Binder Code", this.selectedBinder)
   }
+  /**
+ * Loads and updates binder details for a specific binder identified by the provided code.
+ * Retrieves binder data via an HTTP request and updates the 'binderListDetails' property.
+ * @param code - The code of the binder for which details are loaded.
+ */
   loadBinders(code:any){
     return this.binderService.getBinders(code).subscribe(res=>{
       this.binderListDetails=res;
       this.cdr.detectChanges();
     });
   }
+  /**
+ * Retrieves premium data via an HTTP request using the provided section code, binder code, and subclass code.
+ * It then updates the 'premiumList' with the retrieved data and filters 'selectedPremiumList' to include
+ * premiums that match the given codes. Finally, it triggers change detection to update the UI.
+ * @param binderCode - The code of the binder for which premiums are loaded.
+ * @param subclassCode - The code of the subclass for which premiums are loaded.
+ */
   loadAllPremiums( binderCode:any, subclassCode:any){
    let sectionCode =  this.selectedSection.sectionCode 
     // log.debug(sectionCode,binderCode,subclassCode, "The code from get all premiums")
@@ -176,7 +224,13 @@ export class PremiumRateComponent implements OnInit {
       this.cdr.detectChanges();
 
     })
-  }
+  } 
+  /**
+ * Loads and updates premium rate information for a specific premium identified by the provided code.
+ * Retrieves premium data via an HTTP request using the provided premium code, updates the 'premiumList'
+ * property, sets the 'selectedPremium' to the provided code.
+ * @param code - The code of the premium for which details are loaded.
+ */
   loadPremiums(code:any){
     return this.service.getPremiums(code).subscribe(res=>{
       this.premiumList=res;
@@ -184,6 +238,11 @@ export class PremiumRateComponent implements OnInit {
       this.cdr.detectChanges();
     });
   }
+  /**
+ * Creates and initializes the premium rate form using Angular's FormBuilder.
+ * This method defines form controls and validators for various premium rate attributes,
+ * It also initializes a separate search form for filtering premium rates.
+ */
   createPremiumForm(){
     this.premiumForm=this.fb.group({
       sectionShortDescription: ['', Validators.required],
@@ -218,6 +277,14 @@ export class PremiumRateComponent implements OnInit {
     search:['']
   })
 }
+
+/**
+ * Creates a new premium rate based on the data entered in the premium rate form.
+ * It extracts the form values, sets specific properties like section code, subclass code, and binder code,
+ * and sends the data to the 'service' for creating a new premium rate via an HTTP request.
+ * After submission, it displays a success message if successful or an error message in case of failure,
+ * and resets the premium rate form.
+ */
 createPremium(){
   const premium = this.premiumForm.value
   premium.sectionCode=this.selectedSection.sectionCode;
@@ -238,6 +305,12 @@ createPremium(){
   });
   
 }
+/**
+ * Performs a test operation to check if a binder is selected.
+ * If a binder is not selected, it displays an error message using the 'messageService.'
+ * Otherwise, it triggers a click event on an element with the ID 'openModalButton.'
+ * This function is typically used to prompt the user to select a binder before proceeding.
+ */
 test(){
   if(!this.selectedBinder){
     this.messageService.add({severity: 'error', summary: 'Error', detail: 'Select a Binder to continue'});
@@ -246,6 +319,12 @@ test(){
 
   }
 }
+/**
+ * Performs a test operation to check if both a binder and a premium rate are selected for editing.
+ * If either the binder or premium rate is not selected, it displays an error message using the 'messageService.'
+ * Otherwise, it triggers a click event on an element with the ID 'openModalButtonEdit.'
+ * This function is typically used to prompt the user to select both a binder and a premium rate before proceeding with editing.
+ */
 testEdit(){
   if(!this.selectedBinder && !this.selectedPremiumRate){
     this.messageService.add({severity: 'error', summary: 'Error', detail: 'Select a Binder to continue'});
@@ -254,6 +333,12 @@ testEdit(){
 
   }
 }
+/**
+ * Performs a test operation to check if a premium rate is selected for deletion.
+ * If no premium rate is selected, it displays an error message using the 'messageService.'
+ * Otherwise, it triggers a click event on an element with the ID 'openModalButtonDelete.'
+ * This function is typically used to prompt the user to select a premium rate before proceeding with deletion.
+ */
 testDelete(){
   if(!this.selectedPremiumRate){
     this.messageService.add({severity:'error', summary: 'Error', detail: 'Select a Premium to continue'});
@@ -262,11 +347,22 @@ testDelete(){
 
   }
 }
+/**
+ * Handles the selection of a premium rate from a user interface event.
+ * It updates the 'selectedPremiumRate' property with the selected premium rate data
+ * and patches the premium rate form with the selected premium rate values.
+ * @param event - The event containing the selected premium rate data.
+ */
 onselectPremium(event: any){
   this.selectedPremiumRate = event
   this.premiumForm.patchValue( this.selectedPremiumRate)
-  log.info( this.selectedPremiumRate)
+  // log.info( this.selectedPremiumRate)
 }
+/**
+ * Updates an existing premium rate with new data provided in the premium rate form.
+ * After the update operation, it displays a success message if successful or an error message in case of failure,
+ * and resets the premium rate form.
+ */
 updatePremium(){
   const premium = this.premiumForm.value
   premium.sectionCode=this.selectedSection.sectionCode;
@@ -284,21 +380,51 @@ updatePremium(){
     }  
   })
 }
-deletePremium(){
-  let id = this.selectedPremiumRate.code;
-    this.service.deletePremium(id).subscribe((data)=>{
-      try{
-        this.messageService.add({severity:'success', summary: 'Success', detail: 'Deleted Succesfully'});
-      }catch(error){
-        this.messageService.add({severity:'error', summary: 'Error', detail: 'Error, try again later'});
+// deletePremium(){
+//   let id = this.selectedPremiumRate.code;
+//     this.service.deletePremium(id).subscribe((data)=>{
+//       try{
+//         this.messageService.add({severity:'success', summary: 'Success', detail: 'Deleted Succesfully'});
+//       }catch(error){
+//         this.messageService.add({severity:'error', summary: 'Error', detail: 'Error, try again later'});
+//       }
+//     })
+// }
+/**
+ * Deletes a premium rate based on the selected premium rate and its code.
+ * If a valid premium rate with a code is selected, it sends a delete request to the 'service' to delete the premium rate.
+ * After the deletion operation, it displays a success message if successful or an error message in case of failure.
+ * If no valid premium rate is selected, it displays an error message to prompt the user to select a premium rate for deletion.
+ */
+deletePremium() {
+  if (this.selectedPremiumRate && this.selectedPremiumRate.code) {
+    let id = this.selectedPremiumRate.code;
+    this.service.deletePremium(id).subscribe((data) => {
+      try {
+        this.messageService.add({severity: 'success', summary: 'Success', detail: 'Deleted Successfully'});
+      } catch (error) {
+        this.messageService.add({severity: 'error', summary: 'Error', detail: 'Error, try again later'});
       }
-    })
+    });
+  } else {
+    // Handle the case where selectedPremiumRate is undefined or does not have a code property
+    this.messageService.add({severity: 'error', summary: 'Error', detail: 'Select a Premium to continue'});
+  }
 }
+/**
+ * Applies a global filter to a data table component to filter its contents based on the provided search value.
+ * This method is typically used for global searching across all columns of a data table.
+ * @param $event - The event that triggered the filter action.
+ * @param stringVal - The search value to filter the data table contents.
+ */
 applyFilterGlobal($event, stringVal) {
   log.info(`calling global filter`, stringVal);
   this.dt1.filterGlobal(($event.target as HTMLInputElement).value, stringVal);
 }
-
+/**
+ * Displays the "Add Premium" card by setting the corresponding boolean flags and hiding other related cards and buttons.
+ * This method is typically used to show the form for adding a new premium rate and hide other unrelated elements on the UI.
+ */
 showAddPremiumCard() {
   this.showAddPremium = true;
   this.showEditPremium = false;
@@ -306,11 +432,18 @@ showAddPremiumCard() {
   this.showcover = false;
   this.showButtons = false;
 }
-
+/**
+ * Hides the "Add Premium" card by setting the corresponding boolean flags and showing related buttons.
+ * This method is typically used to hide the form for adding a new premium rate and show other related elements like buttons.
+ */
 hideAddPremiumCard() {
   this.showAddPremium = false;
   this.showButtons = true;
 }
+/**
+ * Displays the "Edit Premium" card by setting the corresponding boolean flags and hiding other related cards and buttons.
+ * This method is typically used to show the form for editing an existing premium rate and hide other unrelated elements on the UI.
+ */
 showEditPremiumCard() {
   this.showAddPremium = false;
   this.showEditPremium = true;
@@ -318,6 +451,10 @@ showEditPremiumCard() {
   this.showcover = false;
   this.showButtons = false;
 }
+/**
+ * Hides the "Edit Premium" card by setting the corresponding boolean flags and showing related buttons.
+ * This method is typically used to hide the form for adding a new premium rate and show other related elements like buttons.
+ */
 hideEditPremiumCard() {
   this.showEditPremium = false;
   this.showButtons = true;
